requirements.txt
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
torch>=2.2.0
transformers>=4.38.0
yfinance>=0.2.50
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
feedparser>=6.0.0
tqdm>=4.66.0
pyyaml>=6.0.0
finnhub-python>=2.4.0
matplotlib>=3.8.0
seaborn>=0.13.0
nbformat>=5.9.0
jupyter>=1.0.0
ipykernel>=6.29.0
